update dbservice to remove cheat methods;
create letterContents dbservice;
move queries that return users into user dbservice, etc;
