+++ src/database/users/UserKey.dbservice.ts
@@ -1,8 +1,9 @@
-import { v4 as uuid } from "uuid";
 import { DatabaseService } from "../dbservice";
 import { UserKey } from "./UserKey.dbmodel";
 
-const userTableName: string = "users";
+const sentLetterTableName = "sent_letters";
+const letterTableName = "letters";
+const userTableName = "users";
 
 export class UserKeyDbService extends DatabaseService<UserKey> {
   constructor() {
@@ -21,7 +22,7 @@ export class UserKeyDbService extends DatabaseService<UserKey> {
 
   /**
    * select userkey by public address
-   * @param publicAddress 
+   * @param publicAddress
    */
   async selectUserKey(publicAddress: string): Promise<UserKey> | null {
     const values = [publicAddress];
@@ -53,6 +54,20 @@ export class UserKeyDbService extends DatabaseService<UserKey> {
     );
   }
 
+  /**
+   * unsent recipient keys
+   * retrieval of all unsent letter recipients keys (UserKeys) rather than full letter history
+   * for a given letter id (for either requestor or writer)
+   * @param letterId
+   */
+  async selectAllUnsentRecipientKeysByLetterId(
+    letterId: string
+  ): Promise<UserKey[]> {
+    const queryText = this.selectAllUnsentRecipientKeyByLetterIdQuery;
+    const values = [letterId];
+    return super.runParameterizedQueryWithValuesArray(queryText, values);
+  }
+
   private selectUserKeyQuery = {
     text:
       "select public_address, name, public_key from " +
@@ -65,6 +80,17 @@ export class UserKeyDbService extends DatabaseService<UserKey> {
       "update " +
       userTableName +
       " set public_key = $1 where public_address = $2;",
+  };
+
+  private selectAllUnsentRecipientKeyByLetterIdQuery = {
+    text:
+      "select distinct W.public_address, W.name, W.public_key from " +
+      letterTableName +
+      " as L inner join " +
+      sentLetterTableName +
+      " as S on L.letter_id = S.letter_id join " +
+      userTableName +
+      " as W on S.letter_recipient = W.public_address where L.letter_id = $1 and S.sent_at is NULL order by W.public_address ASC;",
   };
 
   protected dbRowToDbModel(dbRow: any): UserKey {
